Implemented TextToMorse, MorseTranslator, and MorsePlayer classes, which successfully translates text to Morse by implementing its own tones using Android ToneGenerator. Current commit works but has not gone through extended testing.
TextToMorse.java and activity_text_to_morse were updated to include SeekBars which adjust the pitch and volume of the audio playback.

Also updated are the menu items, including an about section. Eventually a preferences activity will be added as well, which will allow the user to adjust the speed/timing of the audio playback.